# shared/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8090
  description: Generated server url
paths:
...
  /api/stall-reasons:
    get:
      tags:
      - catalogs
      summary: List stall reasons
      operationId: listStallReasons
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogDto"
    post:
      tags:
      - catalogs
      summary: Create stall reason (admin)
      operationId: createStallReason
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCatalogRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CatalogDto"
...
  /api/leads/{id}/activities:
    get:
      tags:
      - activities
      summary: Activities of a lead (ownership enforced)
      operationId: list_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityDto"
    post:
      tags:
      - activities
      summary: Log an activity; contact kinds reset the lead's SLA clock
      operationId: add
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDto"
  /api/courses:
    get:
      tags:
      - catalogs
      summary: List courses
      operationId: listCourses
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogDto"
    post:
      tags:
      - catalogs
      summary: Create course (admin)
      operationId: createCourse
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCatalogRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CatalogDto"
  /api/channels:
    get:
      tags:
      - catalogs
      summary: List channels
      operationId: listChannels
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogDto"
    post:
      tags:
      - catalogs
      summary: Create channel (admin)
      operationId: createChannel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCatalogRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CatalogDto"
...
  /api/activities/{id}/complete:
    post:
      tags:
      - activities
      summary: Mark a follow-up as done
      operationId: complete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDto"
...
  /api/stall-reasons/{id}:
    patch:
      tags:
      - catalogs
      summary: Rename or (de)activate stall reason (admin)
      operationId: updateStallReason
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCatalogRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CatalogDto"
...
  /api/courses/{id}:
    patch:
      tags:
      - catalogs
      summary: Rename or (de)activate course (admin)
      operationId: updateCourse
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCatalogRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CatalogDto"
  /api/channels/{id}:
    patch:
      tags:
      - catalogs
      summary: Rename or (de)activate channel (admin)
      operationId: updateChannel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCatalogRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CatalogDto"
...
  /api/analytics/timeseries:
    get:
      tags:
      - analytics
      summary: Leads created vs converted over time
      operationId: timeseries
      parameters:
      - name: interval
        in: query
        required: false
        schema:
          type: string
          default: week
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimePoint"
  /api/analytics/summary:
    get:
      tags:
      - analytics
      summary: Headline KPIs
      operationId: summary
      parameters:
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Summary"
...
components:
  schemas:
...
    CreateCatalogRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
      required:
      - name
    CatalogDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        active:
          type: boolean
...
    CreateActivityRequest:
      type: object
      properties:
        type:
          type: string
          enum:
          - NOTE
          - CALL
          - EMAIL
          - WHATSAPP
          - MEETING
          - FOLLOW_UP
        content:
          type: string
          minLength: 1
        dueAt:
          type: string
          format: date-time
      required:
      - content
      - type
    ActivityDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        leadId:
          type: string
          format: uuid
        type:
          type: string
          enum:
          - NOTE
          - CALL
          - EMAIL
          - WHATSAPP
          - MEETING
          - FOLLOW_UP
        content:
          type: string
        dueAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        createdByName:
          type: string
        createdAt:
          type: string
          format: date-time
...
    UpdateCatalogRequest:
      type: object
      properties:
        name:
          type: string
        active:
          type: boolean
...
    TimePoint:
      type: object
      properties:
        period:
          type: string
        created:
          type: integer
          format: int64
        converted:
          type: integer
          format: int64
    Summary:
      type: object
      properties:
        totalLeads:
          type: integer
          format: int64
        newLeads:
          type: integer
          format: int64
        conversionRate:
          type: number
          format: double
        avgDaysToConvert:
          type: number
          format: double
        slaBreaches:
          type: integer
          format: int64
...